pyproject.toml
[tool.poetry]
name = "cw-habits-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cw_habits_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.3"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
djangorestframework = "^3.16.0"
sqlparse = "^0.5.3"
python-dotenv = "^1.1.0"
black = "^25.1.0"
djangorestframework-simplejwt = "^5.5.0"
celery = "^5.5.3"
redis = "^6.2.0"
django-celery-beat = "^2.8.1"
eventlet = "^0.40.0"
requests = "^2.32.4"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
